Feature/int64 to int (#37)
* Remove pointer from responses

* Update unit tests

* Update integration tests

* Update examples

* Update customer struct

* Order struct fields

* Change int64 to int

* Switch params order

+++ pkg/oauth/response.go
@@ -7,6 +7,6 @@ type Response struct {
 	RefreshToken string `json:"refresh_token"`
 	PublicKey    string `json:"public_key"`
 	TokenType    string `json:"token_type"`
-	ExpiresIn    int64  `json:"expires_in"`
+	ExpiresIn    int    `json:"expires_in"`
 	LiveMode     bool   `json:"live_mode"`
 }
